Introduce Cardano.DB and Cardano.Pool and add skeleton for Cardano.Pool.Sqlite
revert changes to Cardano.Wallet.DB.Model

fix in Cardano.Wallet.DB.Model

bench db compilation fix
