+++ lib/core/src/Cardano/DB/Sqlite.hs
@@ -1,0 +1,234 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric #-}
+{-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedLabels #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE UndecidableInstances #-}
+
+{-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
+{-# HLINT ignore "Redundant flip" #-}
+
+-- |
+-- Copyright: © 2018-2019 IOHK
+-- License: Apache-2.0
+--
+-- An implementation of the DBLayer which uses Persistent and SQLite.
+
+module Cardano.DB.Sqlite
+    ( SqliteContext (..)
+    , chunkSize
+    , dbChunked
+    , destroyDBLayer
+    , handleConstraint
+    , queryLogFunc
+    , startSqliteBackend
+    , transformTrace
+    , unsafeRunQuery
+    ) where
+
+import Prelude
+
+import Cardano.BM.Data.LogItem
+    ( PrivacyAnnotation (..) )
+import Cardano.BM.Data.Severity
+    ( Severity (..) )
+import Cardano.BM.Observer.Monadic
+    ( bracketObserveIO )
+import Cardano.BM.Trace
+    ( Trace, appendName, logDebug, traceNamedItem )
+import Control.Concurrent.MVar
+    ( newMVar, withMVar )
+import Control.Monad
+    ( mapM_ )
+import Control.Monad.Catch
+    ( MonadCatch (..), handleJust )
+import Control.Monad.IO.Class
+    ( MonadIO (..) )
+import Control.Monad.Logger
+    ( LogLevel (..) )
+import Control.Tracer
+    ( contramap )
+import Data.Aeson
+    ( ToJSON )
+import Data.List.Split
+    ( chunksOf )
+import Data.Maybe
+    ( fromMaybe )
+import Data.Text
+    ( Text )
+import Database.Persist.Sql
+    ( LogFunc, Migration, close', runMigrationSilent, runSqlConn )
+import Database.Persist.Sqlite
+    ( SqlBackend, SqlPersistT, mkSqliteConnectionInfo, wrapConnectionInfo )
+import Database.Sqlite
+    ( Error (ErrorConstraint), SqliteException (SqliteException) )
+import Fmt
+    ( fmt, (+|), (+||), (|+), (||+) )
+import GHC.Generics
+    ( Generic )
+import System.Log.FastLogger
+    ( fromLogStr )
+
+
+import qualified Cardano.BM.Configuration.Model as CM
+import qualified Data.ByteString.Char8 as B8
+import qualified Data.Text as T
+import qualified Data.Text.Encoding as T
+import qualified Database.Sqlite as Sqlite
+
+{-------------------------------------------------------------------------------
+                            Sqlite connection set up
+-------------------------------------------------------------------------------}
+
+-- | Context for the SQLite 'DBLayer'.
+data SqliteContext = SqliteContext
+    { getSqlBackend :: SqlBackend
+    -- ^ A handle to the Persistent SQL backend.
+    , runQuery :: forall a. SqlPersistT IO a -> IO a
+    -- ^ 'safely' run a query with logging and lock-protection
+    , dbFile :: Maybe FilePath
+    -- ^ The actual database file, if any. If none, runs in-memory
+    , trace :: Trace IO DBLog
+    -- ^ A 'Trace' for logging
+    }
+
+-- | Run a raw query from the outside using an instantiate DB layer. This is
+-- completely unsafe because it breaks the abstraction boundary and can have
+-- disastrous results on the database consistency.
+unsafeRunQuery :: SqliteContext -> SqlPersistT IO a -> IO a
+unsafeRunQuery = runQuery
+
+queryLogFunc :: Trace IO DBLog -> LogFunc
+queryLogFunc trace _loc _source level str = dbLog trace (MsgQuery msg sev)
+  where
+    -- Filter out parameters which appear after the statement semicolon.
+    -- They will contain sensitive material that we don't want in the log.
+    stmt = B8.takeWhile (/= ';') $ fromLogStr str
+    msg = T.decodeUtf8 stmt
+    sev = case level of
+        LevelDebug -> Debug
+        LevelInfo -> Info
+        LevelWarn -> Warning
+        LevelError -> Error
+        LevelOther _ -> Warning
+
+-- | Run an action, and convert any Sqlite constraints exception into the given
+-- error result. No other exceptions are handled.
+handleConstraint :: MonadCatch m => e -> m a -> m (Either e a)
+handleConstraint e = handleJust select handler . fmap Right
+  where
+      select (SqliteException ErrorConstraint _ _) = Just ()
+      select _ = Nothing
+      handler = const . pure  . Left $ e
+
+-- | Finalize database statements and close the database connection.
+destroyDBLayer :: SqliteContext -> IO ()
+destroyDBLayer (SqliteContext {getSqlBackend, trace, dbFile}) = do
+    logDebug trace (MsgClosing dbFile)
+    close' getSqlBackend
+
+{-------------------------------------------------------------------------------
+                           Internal / Database Setup
+-------------------------------------------------------------------------------}
+
+-- | Opens the SQLite database connection, sets up query logging and timing,
+-- runs schema migrations if necessary.
+startSqliteBackend
+    :: CM.Configuration
+    -> Migration
+    -> Trace IO DBLog
+    -> Maybe FilePath
+    -> IO SqliteContext
+startSqliteBackend logConfig migrateAll trace fp = do
+    let traceQuery = appendName "query" trace
+    backend <- createSqliteBackend trace fp (queryLogFunc traceQuery)
+    lock <- newMVar ()
+    let observe :: IO a -> IO a
+        observe = bracketObserveIO logConfig traceQuery Debug "query"
+    let runQuery :: SqlPersistT IO a -> IO a
+        runQuery cmd = withMVar lock $ const $ observe $ runSqlConn cmd backend
+    migrations <- runQuery $ runMigrationSilent migrateAll
+    dbLog trace $ MsgMigrations (length migrations)
+    pure $ SqliteContext backend runQuery fp trace
+
+createSqliteBackend
+    :: Trace IO DBLog
+    -> Maybe FilePath
+    -> LogFunc
+    -> IO SqlBackend
+createSqliteBackend trace fp logFunc = do
+    let connStr = sqliteConnStr fp
+    dbLog trace $ MsgConnStr connStr
+    conn <- Sqlite.open connStr
+    wrapConnectionInfo (mkSqliteConnectionInfo connStr) conn logFunc
+
+sqliteConnStr :: Maybe FilePath -> Text
+sqliteConnStr = maybe ":memory:" T.pack
+
+{-------------------------------------------------------------------------------
+                                    Logging
+-------------------------------------------------------------------------------}
+
+data DBLog
+    = MsgMigrations Int
+    | MsgQuery Text Severity
+    | MsgConnStr Text
+    | MsgClosing (Maybe FilePath)
+    deriving (Generic, Show, Eq, ToJSON)
+
+transformTrace :: Trace IO Text -> Trace IO DBLog
+transformTrace = contramap (fmap dbLogText)
+
+dbLog :: MonadIO m => Trace m DBLog -> DBLog -> m ()
+dbLog logTrace msg = traceNamedItem logTrace Public (dbLogLevel msg) msg
+
+dbLogLevel :: DBLog -> Severity
+dbLogLevel (MsgMigrations 0) = Debug
+dbLogLevel (MsgMigrations _) = Notice
+dbLogLevel (MsgQuery _ sev) = sev
+dbLogLevel (MsgConnStr _) = Debug
+dbLogLevel (MsgClosing _) = Debug
+
+dbLogText :: DBLog -> Text
+dbLogText (MsgMigrations 0) = "No database migrations were necessary."
+dbLogText (MsgMigrations n) = fmt $ ""+||n||+" migrations were applied to the database."
+dbLogText (MsgQuery stmt _) = stmt
+dbLogText (MsgConnStr connStr) = "Using connection string: " <> connStr
+dbLogText (MsgClosing fp) = "Closing database ("+|fromMaybe "in-memory" fp|+")"
+
+{-------------------------------------------------------------------------------
+                               Extra DB Helpers
+-------------------------------------------------------------------------------}
+
+-- | Convert a single DB "updateMany" (or similar) query into multiple
+-- updateMany queries with smaller lists of values.
+--
+-- This is to prevent too many variables appearing in the SQL statement.
+-- SQLITE_MAX_VARIABLE_NUMBER is 999 by default, and we will get a
+-- "too many SQL variables" exception if that is exceeded.
+--
+-- We choose a conservative value 'chunkSize' << 999 because there can be
+-- multiple variables per row updated.
+dbChunked :: ([a] -> SqlPersistT IO b) -> [a] -> SqlPersistT IO ()
+dbChunked = chunkedM chunkSize
+
+-- | Given an action which takes a list of items, and a list of items, run that
+-- action multiple times with the input list cut into chunks.
+chunkedM
+    :: Monad m
+    => Int -- ^ Chunk size
+    -> ([a] -> m b) -- ^ Action to run on values
+    -> [a] -- ^ The values
+    -> m ()
+chunkedM n f = mapM_ f . chunksOf n
+
+-- | Size of chunks when inserting, updating or deleting many rows at once. We
+-- only act on `chunkSize` values at a time. See also 'dbChunked'.
+chunkSize :: Int
+chunkSize = 100
