+++ lib/core/src/Cardano/Pool/DB/Sqlite.hs
@@ -1,0 +1,95 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE DuplicateRecordFields #-}
+{-# LANGUAGE FlexibleContexts #-}
+{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE OverloadedLabels #-}
+{-# LANGUAGE RankNTypes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE UndecidableInstances #-}
+
+{-# OPTIONS_GHC -Wno-unrecognised-pragmas #-}
+{-# HLINT ignore "Redundant flip" #-}
+
+-- |
+-- Copyright: © 2018-2019 IOHK
+-- License: Apache-2.0
+--
+-- An implementation of the DBLayer which uses Persistent and SQLite.
+
+module Cardano.Pool.DB.Sqlite
+    ( newDBLayer
+    , withDBLayer
+    ) where
+
+import Prelude
+
+import Cardano.BM.Trace
+    ( Trace )
+import Cardano.DB.Sqlite
+    ( SqliteContext, destroyDBLayer, startSqliteBackend, transformTrace )
+import Cardano.Pool.DB
+    ( DBLayer (..) )
+import Cardano.Pool.DB.Sqlite.TH
+    ( migrateAll )
+import Control.Exception
+    ( bracket )
+import Data.Text
+    ( Text )
+
+import qualified Cardano.BM.Configuration.Model as CM
+
+-- | Runs an action with a connection to the SQLite database.
+--
+-- Database migrations are run to create tables if necessary.
+--
+-- If the given file path does not exist, it will be created by the sqlite
+-- library.
+withDBLayer
+    :: CM.Configuration
+       -- ^ Logging configuration
+    -> Trace IO Text
+       -- ^ Logging object
+    -> Maybe FilePath
+       -- ^ Database file location, or Nothing for in-memory database
+    -> (DBLayer IO -> IO a)
+       -- ^ Action to run.
+    -> IO a
+withDBLayer logConfig trace fp action =
+    bracket before after (action . snd)
+  where
+    before = newDBLayer logConfig trace fp
+    after = destroyDBLayer . fst
+
+-- | Sets up a connection to the SQLite database.
+--
+-- Database migrations are run to create tables if necessary.
+--
+-- If the given file path does not exist, it will be created by the sqlite
+-- library.
+--
+-- 'getDBLayer' will provide the actual 'DBLayer' implementation. The database
+-- should be closed with 'destroyDBLayer'. If you use 'withDBLayer' then both of
+-- these things will be handled for you.
+newDBLayer
+    :: CM.Configuration
+       -- ^ Logging configuration
+    -> Trace IO Text
+       -- ^ Logging object
+    -> Maybe FilePath
+       -- ^ Database file location, or Nothing for in-memory database
+    -> IO (SqliteContext, DBLayer IO)
+newDBLayer logConfig trace fp = do
+    let trace' = transformTrace trace
+    ctx <- startSqliteBackend logConfig migrateAll trace' fp
+    return (ctx, DBLayer
+        { putPoolProduction = undefined
+
+        , readPoolProduction = undefined
+
+        , rollbackTo = undefined
+
+        , cleanDB = undefined
+
+        })
