+++ lib/core/src/Cardano/Wallet/DB/Sqlite.hs
@@ -1,7 +1,5 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DuplicateRecordFields #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE LambdaCase #-}
@@ -24,13 +22,10 @@
 -- An implementation of the DBLayer which uses Persistent and SQLite.
 
 module Cardano.Wallet.DB.Sqlite
-    ( SqliteContext
-    , newDBLayer
-    , destroyDBLayer
+    ( newDBLayer
     , mkDBFactory
     , findDatabases
     , withDBLayer
-    , unsafeRunQuery
 
     -- * Interfaces
     , PersistState (..)
@@ -39,16 +34,19 @@ module Cardano.Wallet.DB.Sqlite
 
 import Prelude
 
-import Cardano.BM.Data.LogItem
-    ( PrivacyAnnotation (..) )
-import Cardano.BM.Data.Severity
-    ( Severity (..) )
-import Cardano.BM.Observer.Monadic
-    ( bracketObserveIO )
 import Cardano.BM.Trace
-    ( Trace, appendName, logDebug, logInfo, logNotice, traceNamedItem )
+    ( Trace, appendName, logInfo, logNotice )
 import Cardano.Crypto.Wallet
     ( XPrv )
+import Cardano.DB.Sqlite
+    ( SqliteContext (..)
+    , chunkSize
+    , dbChunked
+    , destroyDBLayer
+    , handleConstraint
+    , startSqliteBackend
+    , transformTrace
+    )
 import Cardano.Wallet.DB
     ( DBFactory (..)
     , DBLayer (..)
@@ -94,22 +92,14 @@ import Control.Exception
     ( bracket )
 import Control.Monad
     ( forM, mapM_, void, when )
-import Control.Monad.Catch
-    ( MonadCatch (..), handleJust )
 import Control.Monad.IO.Class
     ( MonadIO (..) )
-import Control.Monad.Logger
-    ( LogLevel (..) )
 import Control.Monad.Trans.Class
     ( lift )
 import Control.Monad.Trans.Except
     ( ExceptT (..), runExceptT )
 import Control.Monad.Trans.Maybe
     ( MaybeT (..) )
-import Control.Tracer
-    ( contramap )
-import Data.Aeson
-    ( ToJSON )
 import Data.Coerce
     ( coerce )
 import Data.Either
@@ -121,7 +111,7 @@ import Data.List.Split
 import Data.Map.Strict
     ( Map )
 import Data.Maybe
-    ( catMaybes, fromMaybe )
+    ( catMaybes )
 import Data.Quantity
     ( Quantity (..) )
 import Data.Text
@@ -135,10 +125,8 @@ import Data.Word
 import Database.Persist.Sql
     ( Entity (..)
     , Filter
-    , LogFunc
     , SelectOpt (..)
     , Update (..)
-    , close'
     , deleteCascadeWhere
     , deleteWhere
     , insertMany_
@@ -146,8 +134,6 @@ import Database.Persist.Sql
     , rawExecute
     , repsert
     , repsertMany
-    , runMigrationSilent
-    , runSqlConn
     , selectFirst
     , selectKeysList
     , selectList
@@ -161,19 +147,11 @@ import Database.Persist.Sql
     , (>=.)
     )
 import Database.Persist.Sqlite
-    ( SqlBackend, SqlPersistT, mkSqliteConnectionInfo, wrapConnectionInfo )
-import Database.Sqlite
-    ( Error (ErrorConstraint), SqliteException (SqliteException) )
-import Fmt
-    ( fmt, (+|), (+||), (|+), (||+) )
-import GHC.Generics
-    ( Generic )
+    ( SqlPersistT )
 import System.Directory
     ( doesFileExist, listDirectory, removePathForcibly )
 import System.FilePath
     ( (</>) )
-import System.Log.FastLogger
-    ( fromLogStr )
 
 import qualified Cardano.BM.Configuration.Model as CM
 import qualified Cardano.Wallet.Primitive.AddressDerivation as W
@@ -183,62 +161,8 @@ import qualified Cardano.Wallet.Primitive.AddressDiscovery.Random as Rnd
 import qualified Cardano.Wallet.Primitive.AddressDiscovery.Sequential as Seq
 import qualified Cardano.Wallet.Primitive.Model as W
 import qualified Cardano.Wallet.Primitive.Types as W
-import qualified Data.ByteString.Char8 as B8
 import qualified Data.Map as Map
 import qualified Data.Text as T
-import qualified Data.Text.Encoding as T
-import qualified Database.Sqlite as Sqlite
-
-{-------------------------------------------------------------------------------
-                            Sqlite connection set up
--------------------------------------------------------------------------------}
-
--- | Context for the SQLite 'DBLayer'.
-data SqliteContext = SqliteContext
-    { getSqlBackend :: SqlBackend
-    -- ^ A handle to the Persistent SQL backend.
-    , runQuery :: forall a. SqlPersistT IO a -> IO a
-    -- ^ 'safely' run a query with logging and lock-protection
-    , dbFile :: Maybe FilePath
-    -- ^ The actual database file, if any. If none, runs in-memory
-    , trace :: Trace IO DBLog
-    -- ^ A 'Trace' for logging
-    }
-
--- | Run a raw query from the outside using an instantiate DB layer. This is
--- completely unsafe because it breaks the abstraction boundary and can have
--- disastrous results on the database consistency.
-unsafeRunQuery :: SqliteContext -> SqlPersistT IO a -> IO a
-unsafeRunQuery = runQuery
-
-queryLogFunc :: Trace IO DBLog -> LogFunc
-queryLogFunc trace _loc _source level str = dbLog trace (MsgQuery msg sev)
-  where
-    -- Filter out parameters which appear after the statement semicolon.
-    -- They will contain sensitive material that we don't want in the log.
-    stmt = B8.takeWhile (/= ';') $ fromLogStr str
-    msg = T.decodeUtf8 stmt
-    sev = case level of
-        LevelDebug -> Debug
-        LevelInfo -> Info
-        LevelWarn -> Warning
-        LevelError -> Error
-        LevelOther _ -> Warning
-
--- | Run an action, and convert any Sqlite constraints exception into the given
--- error result. No other exceptions are handled.
-handleConstraint :: MonadCatch m => e -> m a -> m (Either e a)
-handleConstraint e = handleJust select handler . fmap Right
-  where
-      select (SqliteException ErrorConstraint _ _) = Just ()
-      select _ = Nothing
-      handler = const . pure  . Left $ e
-
--- | Finalize database statements and close the database connection.
-destroyDBLayer :: SqliteContext -> IO ()
-destroyDBLayer (SqliteContext {getSqlBackend, trace, dbFile}) = do
-    logDebug trace (MsgClosing dbFile)
-    close' getSqlBackend
 
 -- | Runs an action with a connection to the SQLite database.
 --
@@ -355,7 +279,8 @@ newDBLayer
 newDBLayer logConfig trace fp = do
     lock <- newMVar ()
     let trace' = transformTrace trace
-    ctx@SqliteContext{runQuery} <- startSqliteBackend logConfig trace' fp
+    ctx@SqliteContext{runQuery} <-
+        startSqliteBackend logConfig migrateAll trace' fp
     return (ctx, DBLayer
 
         {-----------------------------------------------------------------------
@@ -490,43 +415,6 @@ newDBLayer logConfig trace fp = do
         , withLock = \action ->
               ExceptT $ withMVar lock $ \() -> runExceptT action
         })
-
-{-------------------------------------------------------------------------------
-                           Internal / Database Setup
--------------------------------------------------------------------------------}
-
--- | Opens the SQLite database connection, sets up query logging and timing,
--- runs schema migrations if necessary.
-startSqliteBackend
-    :: CM.Configuration
-    -> Trace IO DBLog
-    -> Maybe FilePath
-    -> IO SqliteContext
-startSqliteBackend logConfig trace fp = do
-    let traceQuery = appendName "query" trace
-    backend <- createSqliteBackend trace fp (queryLogFunc traceQuery)
-    lock <- newMVar ()
-    let observe :: IO a -> IO a
-        observe = bracketObserveIO logConfig traceQuery Debug "query"
-    let runQuery :: SqlPersistT IO a -> IO a
-        runQuery cmd = withMVar lock $ const $ observe $ runSqlConn cmd backend
-    migrations <- runQuery $ runMigrationSilent migrateAll
-    dbLog trace $ MsgMigrations (length migrations)
-    pure $ SqliteContext backend runQuery fp trace
-
-createSqliteBackend
-    :: Trace IO DBLog
-    -> Maybe FilePath
-    -> LogFunc
-    -> IO SqlBackend
-createSqliteBackend trace fp logFunc = do
-    let connStr = sqliteConnStr fp
-    dbLog trace $ MsgConnStr connStr
-    conn <- Sqlite.open connStr
-    wrapConnectionInfo (mkSqliteConnectionInfo connStr) conn logFunc
-
-sqliteConnStr :: Maybe FilePath -> Text
-sqliteConnStr = maybe ":memory:" T.pack
 
 delegationToPoolId :: W.WalletDelegation W.PoolId -> Maybe W.PoolId
 delegationToPoolId W.NotDelegating = Nothing
@@ -1103,65 +991,3 @@ selectRndStatePending wid = do
 -- return a placeholder.
 selectRndStateKey :: W.WalletId -> SqlPersistT IO (Rnd.RndKey 'RootK XPrv)
 selectRndStateKey wid = maybe Rnd.nullKey fst <$> selectPrivateKey wid
-
-{-------------------------------------------------------------------------------
-                                    Logging
--------------------------------------------------------------------------------}
-
-data DBLog
-    = MsgMigrations Int
-    | MsgQuery Text Severity
-    | MsgConnStr Text
-    | MsgClosing (Maybe FilePath)
-    deriving (Generic, Show, Eq, ToJSON)
-
-transformTrace :: Trace IO Text -> Trace IO DBLog
-transformTrace = contramap (fmap dbLogText)
-
-dbLog :: MonadIO m => Trace m DBLog -> DBLog -> m ()
-dbLog logTrace msg = traceNamedItem logTrace Public (dbLogLevel msg) msg
-
-dbLogLevel :: DBLog -> Severity
-dbLogLevel (MsgMigrations 0) = Debug
-dbLogLevel (MsgMigrations _) = Notice
-dbLogLevel (MsgQuery _ sev) = sev
-dbLogLevel (MsgConnStr _) = Debug
-dbLogLevel (MsgClosing _) = Debug
-
-dbLogText :: DBLog -> Text
-dbLogText (MsgMigrations 0) = "No database migrations were necessary."
-dbLogText (MsgMigrations n) = fmt $ ""+||n||+" migrations were applied to the database."
-dbLogText (MsgQuery stmt _) = stmt
-dbLogText (MsgConnStr connStr) = "Using connection string: " <> connStr
-dbLogText (MsgClosing fp) = "Closing database ("+|fromMaybe "in-memory" fp|+")"
-
-{-------------------------------------------------------------------------------
-                               Extra DB Helpers
--------------------------------------------------------------------------------}
-
--- | Convert a single DB "updateMany" (or similar) query into multiple
--- updateMany queries with smaller lists of values.
---
--- This is to prevent too many variables appearing in the SQL statement.
--- SQLITE_MAX_VARIABLE_NUMBER is 999 by default, and we will get a
--- "too many SQL variables" exception if that is exceeded.
---
--- We choose a conservative value 'chunkSize' << 999 because there can be
--- multiple variables per row updated.
-dbChunked :: ([a] -> SqlPersistT IO b) -> [a] -> SqlPersistT IO ()
-dbChunked = chunkedM chunkSize
-
--- | Given an action which takes a list of items, and a list of items, run that
--- action multiple times with the input list cut into chunks.
-chunkedM
-    :: Monad m
-    => Int -- ^ Chunk size
-    -> ([a] -> m b) -- ^ Action to run on values
-    -> [a] -- ^ The values
-    -> m ()
-chunkedM n f = mapM_ f . chunksOf n
-
--- | Size of chunks when inserting, updating or deleting many rows at once. We
--- only act on `chunkSize` values at a time. See also 'dbChunked'.
-chunkSize :: Int
-chunkSize = 100
